I've made some significant changes to the UI based on your feedback, aiming for an enhanced chat and reporting experience.
Here's a summary of the key changes:

1.  **Three-Column Layout:**
    *   I've introduced a new right sidebar.
    *   The application now uses a three-column layout: Left Sidebar, Main Content, Right Sidebar.

2.  **Main Content Area (Chat Focus):**
    *   I've removed the "Summary" and "Original Content" tabs from the main content area.
    *   The main content area is now dedicated to the chat interface when a notebook is active.

3.  **Right Sidebar (Reporting):**
    *   I've moved the "Generate HTML Report" functionality into the new right sidebar.
    *   You can select sources from the current notebook and generate reports.

4.  **Left Sidebar (Sources & Context):**
    *   I've relocated the "➕ Add YouTube Source" and "➕ Add File Source" buttons to the left sidebar, and they will appear when a notebook is selected.
    *   Clicking a source (YouTube or file) in the left sidebar now toggles its inclusion in the LLM's chat context.
    *   Visual indicators (e.g., '✓'/'☐') show which sources are currently part of the context.
    *   I've removed the previous detailed source view that was triggered by clicking a source.

5.  **State Management & Component Refactoring:**
    *   I've simplified state management related to source selection (`selectedSourceId` has been removed).
    *   I've refactored `App.js`, `MainContent.js`, `Sidebar.js`, and created `RightSidebar.js` to accommodate these changes.
    *   I've updated `App.css` for the new layout and styling of new/modified elements.

These changes are designed to streamline your workflow by providing a dedicated chat area, centralizing source management and context selection in the left sidebar, and dedicating the right sidebar to report generation.
