+++ frontend/src/App.css
@@ -37,7 +37,7 @@ body {
 
 .app-container {
   display: flex;
-  height: 100vh;
+  min-height: 100vh; /* Ensure it takes full viewport height */
   /* background-color: var(--neutral-bg); Rely on index.css or body rule */
 }
 
@@ -99,7 +99,7 @@ input[type="text"]:focus, textarea:focus {
 
 /* Sidebar Enhancements */
 .sidebar {
-  width: 320px; /* Slightly wider */
+  width: 250px; /* Fixed width for left sidebar */
   flex-shrink: 0; /* Prevent sidebar from shrinking */
   border-right: 1px solid var(--neutral-border); /* Softer border */
   padding: 20px;
@@ -505,8 +505,8 @@ button.secondary:hover {
 
 /* Main Content Enhancements */
 .main-content {
-  flex-grow: 1;
-  padding: 25px 30px; /* More padding */
+  flex-grow: 1; /* Takes up remaining space */
+  padding: 1rem;
   overflow-y: auto;
   display: flex;
   flex-direction: column;
@@ -514,6 +514,16 @@ button.secondary:hover {
   border-radius: var(--default-border-radius); /* ADDED */
   margin-left: 10px; /* If sidebar has shadow */
   box-shadow: var(--default-box-shadow); /* ADDED */
+}
+
+.right-sidebar {
+  width: 300px; /* Fixed width for right sidebar */
+  flex-shrink: 0; /* Prevent sidebar from shrinking */
+  padding: 1rem;
+  background-color: #f0f0f0; /* Placeholder background */
+  overflow-y: auto;
+  border-left: 1px solid var(--neutral-border); /* Separator line */
+  box-shadow: var(--default-box-shadow);
 }
 
 .content-above-chat {
@@ -720,4 +730,31 @@ p {
   cursor: not-allowed;
   opacity: 0.65; /* Standard opacity for disabled elements */
   border-color: var(--neutral-border); /* Lighter border for disabled */
+}
+
+/* Styles for source list item selection in Sidebar */
+.source-list-item {
+  /* Ensure other styles for source list items are compatible */
+  cursor: pointer;
+}
+
+.source-list-item.selected-for-chat .source-name-text { /* If using class to bold */
+  font-weight: bold;
+  color: var(--primary-blue); /* Match other selection colors */
+}
+
+.sidebar .source-list li .source-info .selection-indicator {
+  margin-right: 8px; /* Space between indicator and name */
+  font-weight: bold; /* Make ✓/☐ more prominent */
+  color: var(--primary-blue); /* Indicator color for selected */
+}
+
+.source-list-item:not(.selected-for-chat) .source-info .selection-indicator {
+  color: var(--text-secondary); /* Subdued color for unselected indicator */
+}
+
+.sidebar .source-list li .source-info .source-name-text {
+  /* Ensure proper alignment if icon/indicator and text are on the same line */
+  display: inline-block;
+  vertical-align: middle;
 }
